jcr: moved last jobs list into separate file
