+++ core/src/filed/filed.cc
@@ -180,9 +180,7 @@ int main(int argc, char* argv[])
         break;
 
       case 'z': /* switch network debugging on */
-        if (!BnetDump::EvaluateCommandLineArgs(optarg)) {
-          exit(1);
-        }
+        if (!BnetDump::EvaluateCommandLineArgs(optarg)) { exit(1); }
         break;
 
       case '?':
@@ -275,6 +273,7 @@ int main(int argc, char* argv[])
 
   LoadFdPlugins(me->plugin_directory, me->plugin_names);
 
+  InitJcrChain();
   if (!no_signals) {
     StartWatchdog(); /* start watchdog thread */
     if (me->jcr_watchdog_time) {
