+++ core/src/lib/jcr.cc
@@ -50,6 +50,7 @@
 #include "lib/thread_specific_data.h"
 #include "lib/tls_conf.h"
 #include "lib/bsock.h"
+#include "lib/last_jobs_list.h"
 #include "lib/message_queue_item.h"
 #include "lib/watchdog.h"
 
@@ -76,8 +77,6 @@ static void unlock_jcr_chain();
 #endif
 
 int num_jobs_run;
-dlist* last_jobs = nullptr;
-const int max_last_jobs = 10;
 
 static std::vector<std::weak_ptr<JobControlRecord>> job_control_record_cache;
 static dlist* job_control_record_chain = nullptr;
@@ -85,125 +84,12 @@ static int watch_dog_timeout = 0;
 
 static pthread_mutex_t jcr_lock = PTHREAD_MUTEX_INITIALIZER;
 static pthread_mutex_t job_start_mutex = PTHREAD_MUTEX_INITIALIZER;
-static pthread_mutex_t last_jobs_mutex = PTHREAD_MUTEX_INITIALIZER;
 
 static char Job_status[] = "Status Job=%s JobStatus=%d\n";
 
 void LockJobs() { P(job_start_mutex); }
 
 void UnlockJobs() { V(job_start_mutex); }
-
-void InitLastJobsList()
-{
-  JobControlRecord* jcr = nullptr;
-  struct s_last_job* job_entry = nullptr;
-  if (!last_jobs) { last_jobs = new dlist(job_entry, &job_entry->link); }
-  if (!job_control_record_chain) {
-    job_control_record_chain = new dlist(jcr, &jcr->link);
-  }
-}
-
-void TermLastJobsList()
-{
-  if (last_jobs) {
-    LockLastJobsList();
-    while (!last_jobs->empty()) {
-      void* je = last_jobs->first();
-      last_jobs->remove(je);
-      free(je);
-    }
-    delete last_jobs;
-    last_jobs = nullptr;
-    UnlockLastJobsList();
-  }
-  if (job_control_record_chain) {
-    delete job_control_record_chain;
-    job_control_record_chain = nullptr;
-  }
-}
-
-bool ReadLastJobsList(int fd, uint64_t addr)
-{
-  struct s_last_job *je, job;
-  uint32_t num;
-  bool ok = true;
-
-  Dmsg1(100, "read_last_jobs seek to %d\n", (int)addr);
-  if (addr == 0 || lseek(fd, (boffset_t)addr, SEEK_SET) < 0) { return false; }
-  if (read(fd, &num, sizeof(num)) != sizeof(num)) { return false; }
-  Dmsg1(100, "Read num_items=%d\n", num);
-  if (num > 4 * max_last_jobs) { /* sanity check */
-    return false;
-  }
-  LockLastJobsList();
-  for (; num; num--) {
-    if (read(fd, &job, sizeof(job)) != sizeof(job)) {
-      BErrNo be;
-      Pmsg1(000, "Read job entry. ERR=%s\n", be.bstrerror());
-      ok = false;
-      break;
-    }
-    if (job.JobId > 0) {
-      je = (struct s_last_job*)malloc(sizeof(struct s_last_job));
-      memcpy((char*)je, (char*)&job, sizeof(job));
-      if (!last_jobs) { InitLastJobsList(); }
-      last_jobs->append(je);
-      if (last_jobs->size() > max_last_jobs) {
-        je = (struct s_last_job*)last_jobs->first();
-        last_jobs->remove(je);
-        free(je);
-      }
-    }
-  }
-  UnlockLastJobsList();
-  return ok;
-}
-
-uint64_t WriteLastJobsList(int fd, uint64_t addr)
-{
-  struct s_last_job* je;
-  uint32_t num;
-  ssize_t status;
-
-  Dmsg1(100, "write_last_jobs seek to %d\n", (int)addr);
-  if (lseek(fd, (boffset_t)addr, SEEK_SET) < 0) { return 0; }
-  if (last_jobs) {
-    LockLastJobsList();
-    /*
-     * First record is number of entires
-     */
-    num = last_jobs->size();
-    if (write(fd, &num, sizeof(num)) != sizeof(num)) {
-      BErrNo be;
-      Pmsg1(000, "Error writing num_items: ERR=%s\n", be.bstrerror());
-      goto bail_out;
-    }
-    foreach_dlist (je, last_jobs) {
-      if (write(fd, je, sizeof(struct s_last_job)) !=
-          sizeof(struct s_last_job)) {
-        BErrNo be;
-        Pmsg1(000, "Error writing job: ERR=%s\n", be.bstrerror());
-        goto bail_out;
-      }
-    }
-    UnlockLastJobsList();
-  }
-
-  /*
-   * Return current address
-   */
-  status = lseek(fd, 0, SEEK_CUR);
-  if (status < 0) { status = 0; }
-  return status;
-
-bail_out:
-  UnlockLastJobsList();
-  return 0;
-}
-
-void LockLastJobsList() { P(last_jobs_mutex); }
-
-void UnlockLastJobsList() { V(last_jobs_mutex); }
 
 /*
  * Get an ASCII representation of the Operation being performed as an english
@@ -369,9 +255,7 @@ JobControlRecord* new_jcr(int size, JCR_free_HANDLER* daemon_free_jcr)
 
   LockJobs();
   lock_jcr_chain();
-  if (!job_control_record_chain) {
-    job_control_record_chain = new dlist(jcr, &jcr->link);
-  }
+  InitJcrChain();
   job_control_record_chain->append(jcr);
   unlock_jcr_chain();
   UnlockJobs();
@@ -494,52 +378,7 @@ static void JcrCleanup(JobControlRecord* jcr, bool is_destructor_call = false)
 
   Dmsg1(debuglevel, "End job=%d\n", jcr->JobId);
 
-  /*
-   * Keep some statistics
-   */
-  struct s_last_job* je = nullptr;
-
-  switch (jcr->getJobType()) {
-    case JT_BACKUP:
-    case JT_VERIFY:
-    case JT_RESTORE:
-    case JT_MIGRATE:
-    case JT_COPY:
-    case JT_ADMIN:
-      /*
-       * Keep list of last jobs, but not Console where JobId==0
-       */
-      if (jcr->JobId > 0) {
-        LockLastJobsList();
-        num_jobs_run++;
-        je = (struct s_last_job*)malloc(sizeof(struct s_last_job));
-        new (je) s_last_job();
-        je->Errors = jcr->JobErrors;
-        je->JobType = jcr->getJobType();
-        je->JobId = jcr->JobId;
-        je->VolSessionId = jcr->VolSessionId;
-        je->VolSessionTime = jcr->VolSessionTime;
-        bstrncpy(je->Job, jcr->Job, sizeof(je->Job));
-        je->JobFiles = jcr->JobFiles;
-        je->JobBytes = jcr->JobBytes;
-        je->JobStatus = jcr->JobStatus;
-        je->JobLevel = jcr->getJobLevel();
-        je->start_time = jcr->start_time;
-        je->end_time = time(nullptr);
-
-        if (!last_jobs) { InitLastJobsList(); }
-        last_jobs->append(je);
-        if (last_jobs->size() > max_last_jobs) {
-          je = (struct s_last_job*)last_jobs->first();
-          last_jobs->remove(je);
-          free(je);
-        }
-        UnlockLastJobsList();
-      }
-      break;
-    default:
-      break;
-  }
+  SetLastJobsStatistics(jcr);
 
   CloseMsg(jcr); /* close messages for this job */
 
@@ -1186,6 +1025,22 @@ bool InitJcrSubsystem(int timeout)
   RegisterWatchdog(wd);
 
   return true;
+}
+
+void InitJcrChain()
+{
+  JobControlRecord* jcr = nullptr;
+  if (!job_control_record_chain) {
+    job_control_record_chain = new dlist(jcr, &jcr->link);
+  }
+}
+
+void CleanupJcrChain()
+{
+  if (job_control_record_chain) {
+    delete job_control_record_chain;
+    job_control_record_chain = nullptr;
+  }
 }
 
 static void JcrTimeoutCheck(watchdog_t* /* self */)
