+++ core/src/lib/message.cc
@@ -37,6 +37,7 @@
 #include "lib/bsock.h"
 #include "lib/util.h"
 #include "lib/watchdog.h"
+#include "lib/last_jobs_list.h"
 #include "lib/messages_resource.h"
 #include "lib/message_destination_info.h"
 #include "lib/message_queue_item.h"
@@ -219,7 +220,6 @@ void InitMsg(JobControlRecord* jcr,
   int i;
 
   if (jcr == NULL && msg == NULL) {
-    InitLastJobsList();
     /*
      * Setup a daemon key then set invalid jcr
      * Maybe we should give the daemon a jcr???
@@ -522,6 +522,7 @@ void TermMsg()
     catalog_db = NULL;
   }
   TermLastJobsList();
+  CleanupJcrChain();
 }
 
 static inline bool OpenDestFile(JobControlRecord* jcr,
